+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 ### Added
 
 - Added a feature to have context menu option to base64 encode selected text. Works only in JSON files.
+- Send Tokens between imported accounts
 
 ### Changed